Detect ESLint location for init
